Attempt at README doc tests met with errors
Importing README as a doctest is very rudimentary.
It suffers from most of the same issues that rustdoc comments have but is
unable to make use of rustdocs features.
While there are libraries that are supposed to help with they either
contribute nothing or are too complicated for the task in question.
The two examples I found include rust-skeptic & doc-comment.

rust-skeptic requires being ran in a build script, which I consider
unacceptable as this forces it to be used by everything that depends on
this library.

doc-comment meanwhile is just an alias for the code I am already using to
test README & seems to be a macro to eliminate the "hassle" of writing
THREE lines.
It's also depended on by 229 libraries!
Perhaps the best part of doc-comment is the library the blog post noting
that this method exists before dismissing it & explaining how to do the
exact same thing using their library.