deriving tuple from abandoned
It's impossible to instantiate any type in the syn library &
it frustrating for how mild of a task this is.
The best thing I found is to just quote! a code that matches my
requirements then manually step through the tree to get what I want.
This is before I even try to find out how to combine/create types AT
MACRO INVOCATION TIME.